Append static_cast to Token::operator==.

+++ src/token.cpp
@@ -4,9 +4,9 @@
 
 bool token::operator==(const token::Token &rhs, const token::Token &lhs)
 {
-    return rhs.type == lhs.type
-        && rhs.precedence == lhs.precedence
-        && rhs.association == lhs.association;
+    return static_cast<int8_t>(rhs.type) == static_cast<int8_t>(lhs.type)
+        && static_cast<unsigned int>(rhs.precedence) == static_cast<unsigned int>(lhs.precedence)
+        && static_cast<int8_t>(rhs.association) == static_cast<int8_t>(lhs.association);
 }
 
 int token::tokenize(std::istream &stream, std::vector<token::Token> tokens)
